refactor: delete dto 수정
